ADC references for BSE1 and BSE2 are now configured correctly for ADC3 Code regenerated added calls of switch_to_defined_channel() in dev_task The code now works as intended with the Brake Pedal Emulator

+++ nucleo-f767ZI-drev23-ecu-firmware/Core/Inc/ext_drivers/pressTrans.h
@@ -19,7 +19,7 @@ struct pressTrans {
 	short percent;
 
 	void *handle; //ADC handle goes here
-	uint8_t channelNum; //ADC channel number
+	uint32_t channelNum; //ADC channel number
 
 	uint16_t(*read_count)(void *arg); //this is a function pointer in C, kinda weird ain't it.
 };
